Move workflows to .github, add helper actions
As we move to a library of reusable actions, we need to have
workflows that are more adaptable to small changes in the process.
To that end, we want to move as many functions out to individual
actions as possible, so that these are centralized and easily
editable, and can be easily called from various steps.

Signed-off-by: Eric Ball <eball@linuxfoundation.org>